backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0

# Configuration & Validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Auth
python-jose[cryptography]==3.3.0

# AI Integration
google-generativeai>=0.8.0
tenacity>=8.2.3

# Development
pytest==7.4.3
black==23.11.0